@@ -43,8 +43,8 @@
               <groupId>org.apache.maven.plugins</groupId>
               <artifactId>maven-compiler-plugin</artifactId>
               <configuration>
-                  <source>15</source>
-                  <target>15</target>
+                  <source>16</source>
+                  <target>16</target>
               </configuration>
           </plugin>
       </plugins>
